Update selected examples to new API
