+++ examples/is_it_a_plane.jl
@@ -25,7 +25,9 @@ f, a, p = lines(reverse(Proj.geod_path(geod, reverse(jfk)..., reverse(sin)...)).
 
 # The projection will always be centered at wherever the plane is.
 
-# We first 
+# We first create an animation through time, representing times (as hours),
+# relative distances along the path, and altitudes of observation, as 
+# a function of time.  This is done by using the Animations.jl library.
 
 times = [0, 0.5, 17.5, 18]
 distances = [0, 0.05, 0.95, 1]
@@ -37,7 +39,7 @@ altitude_animation = Animation(times, altitudes, sineio())
 # In order to investigate this kind of projection, you can create 
 # a GeoAxis with the projection you want, and then change the
 # altitude to see how the zoom works in real time!
-@time begin
+
 fig = Figure()
 sl = Slider(fig[2, 1], range = exp.(LinRange(log(357860), log(35786000), 30)), startvalue = 35786000)
 satview_projection = lift(sl.value) do alt
@@ -46,12 +48,13 @@ end
 ga = GeoAxis(fig[1, 1]; dest = satview_projection)
 meshimage!(ga, -180..180, -90..90, GeoMakie.earth(), shading = NoShading)
 fig
-end
+
 
 record(fig, "plots/plane.mp4", LinRange(0, 1, 120)) do i
     satview_projection[] = "+proj=geos +h=$(round(Int, at(altitude_animation, i*18))) +lon_0=$(Proj.geod_position_relative(inv_line, i)[2]) +lat_0=$(Proj.geod_position_relative(inv_line, i)[1])"
     yield()
 end
+# ![](plane.mp4)
 
 fig = Figure()
 ga = GeoAxis(fig[1, 1]; dest = "+proj=nsper +h=3000000 +lat_0=-20 +lon_0=145")
