+++ examples/rasters.jl
@@ -1,0 +1,3 @@
+using Rasters, RasterDataSources, ArchGDAL
+using GeoMakie, CairoMakie
+
