+++ examples/rotating_earth.jl
@@ -10,7 +10,8 @@ ga = GeoAxis(
     fig[1, 1],
     dest = destnode,
 )
-surface!(ga, -180 .. 180, -90 .. 90, zeros(128, 128); color=rotr90(GeoMakie.earth()), interpolate = false)
+surface!(ga, -180 .. 180, -90 .. 90, zeros(128, 128); color=rotr90(GeoMakie.earth()), interpolate = false, shading = NoShading)
+xlims!(ga, -89, 89) # TODO: FIXME
 fig
 
 record(fig, "rotating_earth_ortho.mp4"; framerate=30) do io
